+++ lib/Braintree.php
@@ -10,14 +10,16 @@ if (version_compare(PHP_VERSION, '5.4.0', '<')) {
     throw new Braintree_Exception('PHP version >= 5.4.0 required');
 }
 
+class Braintree {
 
-function requireDependencies() {
-    $requiredExtensions = ['xmlwriter', 'openssl', 'dom', 'hash', 'curl'];
-    foreach ($requiredExtensions AS $ext) {
-        if (!extension_loaded($ext)) {
-            throw new Braintree_Exception('The Braintree library requires the ' . $ext . ' extension.');
+    public static function requireDependencies() {
+        $requiredExtensions = ['xmlwriter', 'openssl', 'dom', 'hash', 'curl'];
+        foreach ($requiredExtensions AS $ext) {
+            if (!extension_loaded($ext)) {
+                throw new Braintree_Exception('The Braintree library requires the ' . $ext . ' extension.');
+            }
         }
     }
 }
 
-requireDependencies();
+Braintree::requireDependencies();