+++ lib/Braintree/Customer.php
@@ -22,6 +22,8 @@ namespace Braintree;
  * @property-read array  $androidPayCards
  * @property-read array  $amexExpressCheckoutCards
  * @property-read array  $venmoAccounts
+ * @property-read array  $visaCheckoutCards
+ * @property-read array  $masterpassCards
  * @property-read array  $coinbaseAccounts
  * @property-read array  $customFields custom fields passed with the request
  * @property-read string $email
@@ -295,6 +297,22 @@ class Customer extends Base
         }
         $this->_set('venmoAccounts', $venmoAccountArray);
 
+        $visaCheckoutCardArray = [];
+        if (isset($customerAttribs['visaCheckoutCards'])) {
+            foreach ($customerAttribs['visaCheckoutCards'] AS $visaCheckoutCard) {
+                $visaCheckoutCardArray[] = VisaCheckoutCard::factory($visaCheckoutCard);
+            }
+        }
+        $this->_set('visaCheckoutCards', $visaCheckoutCardArray);
+
+        $masterpassCardArray = [];
+        if (isset($customerAttribs['masterpassCards'])) {
+            foreach ($customerAttribs['masterpassCards'] AS $masterpassCard) {
+                $masterpassCardArray[] = MasterpassCard::factory($masterpassCard);
+            }
+        }
+        $this->_set('masterpassCards', $masterpassCardArray);
+
         $usBankAccountArray = array();
         if (isset($customerAttribs['usBankAccounts'])) {
             foreach ($customerAttribs['usBankAccounts'] AS $usBankAccount) {
@@ -311,6 +329,8 @@ class Customer extends Base
             $this->androidPayCards,
             $this->amexExpressCheckoutCards,
             $this->venmoAccounts,
+            $this->visaCheckoutCards,
+            $this->masterpassCards,
             $this->usBankAccounts
         ));
     }