+++ lib/Braintree/PaymentMethodGateway.php
@@ -71,6 +71,10 @@ class PaymentMethodGateway
                 return UsBankAccount::factory($response['usBankAccount']);
             } else if (isset($response['venmoAccount'])) {
                 return VenmoAccount::factory($response['venmoAccount']);
+            } else if (isset($response['visaCheckoutCard'])) {
+                return VisaCheckoutCard::factory($response['visaCheckoutCard']);
+            } else if (isset($response['masterpassCard'])) {
+                return MasterpassCard::factory($response['masterpassCard']);
             } else if (is_array($response)) {
                 return UnknownPaymentMethod::factory($response);
             }
@@ -134,7 +138,13 @@ class PaymentMethodGateway
             'verificationMerchantAccountId',
             'verifyCard',
             'verificationAmount',
-            ['paypal' => ['payee_email']],
+            ['paypal' => [
+                'payee_email',
+                'order_id',
+                'custom_field',
+                'description',
+                'amount',
+            ]],
         ];
         return [
             'billingAddressId',
@@ -285,6 +295,16 @@ class PaymentMethodGateway
         } else if (isset($response['venmoAccount'])) {
             return new Result\Successful(
                 VenmoAccount::factory($response['venmoAccount']),
+                "paymentMethod"
+            );
+        } else if (isset($response['visaCheckoutCard'])) {
+            return new Result\Successful(
+                VisaCheckoutCard::factory($response['visaCheckoutCard']),
+                "paymentMethod"
+            );
+        } else if (isset($response['masterpassCard'])) {
+            return new Result\Successful(
+                MasterpassCard::factory($response['masterpassCard']),
                 "paymentMethod"
             );
         } else if (isset($response['paymentMethodNonce'])) {